refactor: move custom actions from modules to Game service
