+++ src/components/GatheringGameRow.vue
@@ -64,28 +64,8 @@ export default {
       return userId === game.owner;
     });
 
-    const join = async () => {
-      try {
-        await client
-          .service(`game/${game.id}/players`)
-          .create({});
-      } catch (e) {
-        console.log(e);
-      }
-    };
-
-    const leave = async () => {
-      const userId = $store.state.auth.user.id;
-      const player = game.players.find((p) => p.user === userId);
-      if (!player) return;
-      try {
-        await client
-          .service(`game/${game.id}/players`)
-          .remove(player.id);
-      } catch (e) {
-        console.log(e);
-      }
-    };
+    const join = async () => $store.dispatch('game/join', { game });
+    const leave = async () => $store.dispatch('game/leave', { game });
 
     const launch = async () => {
       try {
