+++ src/store/index.js
@@ -2,7 +2,6 @@ import Vue from 'vue';
 import Vuex from 'vuex';
 
 import { FeathersVuex } from '../feathers';
-import gamePlayers from './modules/gamePlayers';
 
 Vue.use(Vuex);
 Vue.use(FeathersVuex);
@@ -27,15 +26,6 @@ export default new Vuex.Store({
   actions: {
   },
   modules: {
-    gamePlayers,
-    gameStatus: {
-      state: {},
-      actions: {},
-    },
-    gameState: {
-      state: {},
-      actions: {},
-    },
   },
   plugins: services,
 });
